requirements.txt
gradio==4.31.5
python-dotenv
streamlit
google-generativeai
